Refactor design to achieve better abstraction
Scraping is implemented and working for the famous painters layout

The SerpScraper class is being built with two choices for target layout, in order to keep an abstract design in mind. SerpScraper is initialized with the response bytes content and an enum as params; the enum configures the SerpScraper class to target the desired results page layout by setting the correct target scraper function and the correct item tag identifiers.

At first blush, the output file looks good but running diff reveals significant conflicts with files/expected_array.json